refactor: extract renderTUI into src/cli/tui/render-tui.ts
Break the circular import between cli.ts and command handlers by moving
renderTUI, RenderTUIOptions, and alt-screen helpers into their own module.
Command handlers now import from tui/render-tui instead of ../../cli.
