+++ src/cli/tui/index.ts
@@ -1,5 +1,6 @@
 export { App } from './App';
 export * from './components';
 export * from './hooks';
+export * from './render';
 export * from './screens';
 export * from './utils';
